Add redirects for cuml-accel docs
We added client-side redirects in the cuml docs, but these were done
incorrectly (they assumed our docs were built using the `html` builder,
not the `dirhtml` builder). I've pushed a PR up to fix this in the docs
source, but to fix the last release we'll want to (at least temporarily)
add redirects here as well.